custom_mcp_host/requirements.txt
botbuilder_core==4.17.0
botbuilder_schema==4.17.0
fastapi==0.116.1
mcp==1.12.2
openai==1.97.1
pydantic==2.11.7
python-dotenv==1.1.1
